requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
sqlalchemy==2.0.36
alembic==1.13.2
databases==0.9.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
requests==2.32.3
tenacity>=8.1.0,<9

python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
streamlit==1.38.0
pytest-json-report==1.5.0
flake8==7.1.1